Add `library.properties` to path filter of "Compile Examples" workflow
For the sake of efficiency, the "Compile Examples" GitHub Actions workflow is configured to only run when relevant files
are modified.

Changes to the `library.properties` metadata file can affect the compilation of the library, so the workflow must be
triggered any time that file is modified. That was not done with the previous paths filter configuration.