Use FQBN as job name in "Compile Examples" workflow
The `jobs.<job_id>.name` key sets the human identifier for the workflow job shown in the workflow run logs.

By default, the name of matrix jobs is based on the job ID combined with the `jobs.<job_id>.strategy.matrix` object,
which can result in an unnecessarily long and cryptic job name. The job ID + the FQBN of the job's target board is
sufficient. The use of this job name format will improve the readability of the workflow run logs.