Fail "Compile Examples" workflow run if report file missing
The `arduino/compile-sketches` action is configured to produce a file containing information about the compilations.
This file is uploaded to a workflow artifact via the `actions/upload-artifact` action and later consumed by the
`arduino/report-size-deltas` action.

By default the `actions/upload-artifact` action step will pass even if the file specified for upload is not found. In
this case where the file will always be present if the workflow is functioning correctly, that behavior could mask a
serious problem with the workflow. It is better to configure the `actions/upload-artifact` action to fail the workflow
run if the file is not found.