Standardize license file
Standardization in license documentation is important because, in addition to making it easy for humans to find this
vital information, it allows machines to automate the process of license type determination, which is useful both for
discovering suitable open source projects as well as checking open source license compliance.

The open source license of this project is already stored in a standardized location in a dedicated license file.
However, even though the project is licensed under a standardized open source license, additional text was added to the
license file which offers the option to purchase an exception for proprietary use of the project. Even though this offer
does not have any legal effect on the open source license, the presence of that text made it so that the open license
type could not be identified with 100% confidence by machines (e.g., the Licensee Gem used by the GitHub website), which
meant identification could only be made by a human carefully evaluating the license text.

Since there is no need to place the exception offer in the license file, it can be moved to the readme, with the license
file containing only the verbatim standardized open source license text. The result is that the project's open source
license type can now be automatically identified, without making any change to the licensing.